+++ src/mxmake/testing/__init__.py
@@ -41,6 +41,7 @@ class template_directory:
             tempdir = tempfile.mkdtemp()
             os.environ["MXMAKE_FILES"] = tempdir
             os.environ["MXMAKE_MXENV_PATH"] = tempdir
+            os.environ["MXMAKE_GH_ACTIONS_PATH"] = tempdir
             try:
                 if self.reset_registry:
                     with reset_template_registry():
@@ -51,6 +52,7 @@ class template_directory:
                 shutil.rmtree(tempdir)
                 del os.environ["MXMAKE_FILES"]
                 del os.environ["MXMAKE_MXENV_PATH"]
+                del os.environ["MXMAKE_GH_ACTIONS_PATH"]
 
         return wrapper